+++ controllers/team.js
@@ -1,3 +1,4 @@
+const { isObjectIdOrHexString } = require("mongoose");
 const { INVALID, NOT_FOUND, OK } = require("../constants/constants");
 const { TEAM_ADDED, UNEXPECTED_ERROR, TEAM_UPDATED, TEAM_MEMBERS_EXCEED, TEAM_DELETED } = require("../constants/messages");
 const TeamModel = require("../models/Team");
@@ -27,6 +28,7 @@ module.exports.update = async (req, res) => {
     let size = req.body?.size || "";
     let member_id = req.body?.member_id || "";
     if(!id) return res.status(NOT_FOUND).send(makeResponse(UNEXPECTED_ERROR));
+    if(!isObjectIdOrHexString(id)) return res.status(NOT_FOUND).send(makeResponse("Invalid ID"));
     if(!name) return res.status(INVALID).send(makeResponse("Name is empty"));
     if(!size) return res.status(INVALID).send(makeResponse("Size is empty"));
     if(!member_id) return res.status(INVALID).send(makeResponse(UNEXPECTED_ERROR));
@@ -45,10 +47,11 @@ module.exports.update = async (req, res) => {
 
 module.exports.delete = async (req, res) => {
     if(!req.auth?.auth) return res.status(UNAUTHORIZED).send(makeResponse(AUTH_FAILED));
-
+    
     let id = req.params?.id || "";
     if(!id) return res.status(NOT_FOUND).send(makeResponse(UNEXPECTED_ERROR));
-
+    if(!isObjectIdOrHexString(id)) return res.status(NOT_FOUND).send(makeResponse("Invalid ID"));
+    
     let value = await TeamModel.deleteOne({_id: id}).catch((e) => {console.log(e)});
     if(value.deletedCount) return res.status(OK).send(makeResponse(TEAM_DELETED));
     return res.status(NOT_FOUND).send(makeResponse(UNEXPECTED_ERROR));
@@ -59,9 +62,11 @@ module.exports.show = async (req, res) => {
 
     let id = req.params?.id || "";
     if(!id) return res.status(NOT_FOUND).send(makeResponse(UNEXPECTED_ERROR));
+    if(!isObjectIdOrHexString(id)) return res.status(NOT_FOUND).send(makeResponse("Invalid ID"));
 
     let value = await TeamModel.findOne({_id: id}).catch((e) => {console.log(e)});
-    return res.status(OK).send(makeResponse("", value));
+    if(value) return res.status(OK).send(makeResponse("", value));
+    return res.status(NOT_FOUND).send(makeResponse("No data found"));
 }
 
 module.exports.list = async (req, res) => {
@@ -78,7 +83,7 @@ module.exports.list = async (req, res) => {
 module.exports.search = async (req, res) => {
     if(!req.auth?.auth) return res.status(UNAUTHORIZED).send(makeResponse(AUTH_FAILED));
 
-    let filter = req.body.filter.toLowerCase() || "";
+    let filter = req.body?.filter?.toLowerCase() || "";
     let fields = ["name"];
     return res.status(OK).send(makeResponse("", {searchedData: await searchData(TeamModel, filter, fields)}));
 }
@@ -87,9 +92,5 @@ module.exports.downloadExcel = async (req, res) => {
     if(!req.auth?.auth) return res.status(UNAUTHORIZED).send(makeResponse(AUTH_FAILED));
 
     let value = await TeamModel.find({}).catch((e) => {console.log(e)});
-    if(!value) return res.status(OK).send(makeResponse(NO_DATA));
-
-    let response = await writeExcelFile(value, ["name", "size"]);
-    if(response) return res.status(OK).send(makeResponse("File written successfully", response));
-    return res.status(NOT_FOUND).send(makeResponse(UNEXPECTED_ERROR));
+    return res.status(OK).send(makeResponse("File written successfully", await writeExcelFile(value, ["name", "size"])));
 }
